requirements.txt
PyQt6>=6.6.0
PyYAML>=6.0.1
dbus-python>=1.3.2
